Update GitHub workflows and sbt configuration
This commit modifies the GitHub workflows file to reference the 'build_jar' command instead of 'build'. Additionally, the sbt configuration is updated to set the 'SimpleTestServer' as the main class and implement 'sbt-assembly' for jar generation. As a result, the 'sbt assembly' command will be used to run tests in the CI process.